+++ packages/material-ui/src/MuiForm5/MuiForm5.tsx
@@ -1,8 +1,8 @@
-import { StatelessComponent } from 'react';
+import { FunctionComponent } from 'react';
 import { withTheme, FormProps } from '@rjsf/core';
 
 import Theme5 from '../Theme5';
 
-const MuiForm5: React.ComponentClass<FormProps<any>> | StatelessComponent<FormProps<any>> = withTheme(Theme5);
+const MuiForm5: React.ComponentClass<FormProps<any>> | FunctionComponent<FormProps<any>> = withTheme(Theme5);
 
 export default MuiForm5;